+++ changelog.md
@@ -1,3 +1,8 @@
+## v1.1.0
+- Port to Geode v5 / Geometry Dash 2.2081
+- Added bugfix for horizontal scrolling on Windows
+- Improve compatibility with most mods by checking for active scroll layers
+
 ## v1.0.1
 - Better compatibility with HJfod's [Quick Volume Controls](mod:hjfod.quick-volume-controls) mod.